tested new validation annotations

+++ m12-lesson1-start/um-webapp/src/main/java/com/baeldung/um/web/dto/UserDTO.java
@@ -31,7 +31,7 @@ public class UserDTO {
 
     private Set<@Email String> alternativeEmailAddresses;
 
-    private Optional<@Positive Long> age = Optional.empty();
+    private Optional<@Positive Integer> age = Optional.empty();
 
     private Set<Role> roles;
 }
